requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
huggingface_hub
streamlit
transformers
sentence-transformers
chromadb